refactor: state files
