+++ src/handlers/solo.rs
@@ -28,7 +28,7 @@ pub fn handle_solo_page_events(app: &mut App, key_event: KeyEvent) {
                 app.game.logic.navigate_history_previous();
             }
         }
-        KeyCode::Char('t' | 'T') if app.lichess_state.puzzle_game.is_some() && app.current_popup.is_none() => {
+        KeyCode::Char('t' | 'T') if app.lichess_state.puzzle_game.is_some() && app.ui_state.current_popup.is_none() => {
             // Show hint in puzzle mode (only when no popup is active)
             app.show_puzzle_hint();
         }
